test: Move list_available_backups tests to view
