+++ app/tests/management/commands/test_restore_db.py
@@ -5,8 +5,6 @@ from django.conf import settings
 from django.core.management import CommandError, call_command
 from django.test import override_settings
 
-from app.management.commands.restore_db import Command
-
 
 def setup_restore_mocks(mocker):
     """Setup mocks for restore command. Returns dict of all mocks for modification."""
@@ -52,58 +50,6 @@ def setup_restore_mocks(mocker):
 def mock_restore_success(mocker):
     """Fixture for tests that just need successful restore without modifications."""
     return setup_restore_mocks(mocker)
-
-
-def describe_list_available_backups():
-    def returns_sorted_list_newest_first(mocker):
-        mock_subprocess = mocker.patch("app.backups.subprocess")
-        mock_subprocess.run.return_value = Mock(
-            returncode=0,
-            stdout="backup_20241210_020000.sql.gz\nbackup_20241214_020000.sql.gz\nbackup_20241212_020000.sql.gz\n",
-        )
-
-        backups = Command.list_available_backups()
-
-        assert len(backups) == 3
-        assert backups[0]["filename"] == "backup_20241214_020000.sql.gz"
-        assert backups[1]["filename"] == "backup_20241212_020000.sql.gz"
-        assert backups[2]["filename"] == "backup_20241210_020000.sql.gz"
-
-    def returns_empty_list_on_rclone_failure(mocker):
-        mock_subprocess = mocker.patch("app.backups.subprocess")
-        mock_subprocess.run.return_value = Mock(returncode=1, stdout="", stderr="error")
-
-        backups = Command.list_available_backups()
-
-        assert backups == []
-
-    def returns_empty_list_when_no_backups(mocker):
-        mock_subprocess = mocker.patch("app.backups.subprocess")
-        mock_subprocess.run.return_value = Mock(returncode=0, stdout="")
-
-        backups = Command.list_available_backups()
-
-        assert backups == []
-
-    def skips_invalid_filenames(mocker):
-        mock_subprocess = mocker.patch("app.backups.subprocess")
-        mock_subprocess.run.return_value = Mock(
-            returncode=0,
-            stdout="backup_20241214_020000.sql.gz\nsome_other_file.txt\nbackup_invalid.sql.gz\n",
-        )
-
-        backups = Command.list_available_backups()
-
-        assert len(backups) == 1
-        assert backups[0]["filename"] == "backup_20241214_020000.sql.gz"
-
-    def includes_display_date(mocker):
-        mock_subprocess = mocker.patch("app.backups.subprocess")
-        mock_subprocess.run.return_value = Mock(returncode=0, stdout="backup_20241214_140000.sql.gz\n")
-
-        backups = Command.list_available_backups()
-
-        assert backups[0]["display_date"] == "December 14, 2024 at 02:00 PM"
 
 
 def describe_backup_validation():
